requirements_windows.txt
anyio==3.6.2
appdirs==1.4.4
apsw==3.41.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
attrs==21.2.0
autopep8==2.0.2
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.0.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.3
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==4.4.2
defer==1.0.4
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
dnspython==2.1.0
docopt==0.6.2
fasteners==0.14.1
fastjsonschema==2.16.3
feedparser==6.0.8
filelock==3.6.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
greenlet==1.1.2
html2text==2020.1.16
html5lib==1.1
httplib2==0.20.2
idna==2.8
ifaddr==0.1.7
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==7.31.1
ipython-genutils==0.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.8.0
lz4==3.1.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mechanize==0.4.7
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
natsort==8.0.2
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.22.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
py7zr==0.11.3
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.11.0
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyRFC3339==1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.25.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Routes==2.5.1
scour==0.38.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.4
sympy==1.9
terminado==0.17.1
texttable==1.6.4
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==1.26.5
virtualenv==20.13.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.3
xdg==5.0.0
zeroconf==0.38.3
zipp==1.0.0
